Implement common functionality class for 2d/3d viewer.
